refactor(all): Refactor to use ES6 Synatx + Modules, Add Tooling
Refactored entire project.

Instead of using a single class object in the global namespace, each method has been moved into it's own file for sake of readability, maintainability, and testability. Some of the original organization has been used to create the new directory structure. For each part of the application (game, debug, graphics, ui), there is a state container class, plus a file for each method.

These can all be consumed bet the ES6 import syntax, example:

`import game, { getMapCell, drawBuildingTile } from "@/app/game"

Will import the game state object, and the two methods `getMapCell` and `drawBuildingTile`. These can then be useded in any other module just as you would use any object or function normally.

State containers are singletons, meaning if a property is changed in one place, every other place where that depends on that state will be updated.

Most methods have been significantly refactored to make them more concise, follow coding best-practices, and to clean up numerous errors in the original source code. Destructuring is heavily used to keep code concise and DRY. In as many places as possible, ternary expressions are used in place of `if/else` blocks for variable assignment, coupled with the use of `const` to prevent accidental re-assignment. Where possible, variable declaration is skipped altogether to improve performance. Many loops have been refactored to use the `for...of` syntax for brevity. Where applicable, object and array literal syntax is used in conjunction with `...rest` to assign default values and overide them in a single assignment instead of each property individually. Default parameters are used in place of if + typeof blocks.

Some tooling has been added but not completely configured. Notable additions are babel, eslint, and prettier. Webpack, codecov, travisci, ava, and wallaby are all forthcoming.

TODO:

Because of the way the original project was setup, there's still some hiccups to getting the script to run properly in electron. Right now it appears the code runs in electron's main process and not the renderer, so the next commit will be focused on ensuring that works.